config.c cleanup: move attribute registration to attributes.c
